+++ src/components/Wiki/Editor.js
@@ -1,6 +1,6 @@
 // Copyright 2023 Paion Data. All rights reserved.
 import React from "react";
-import ExampleTheme from "./themes/ExampleTheme";
+import DefaultTheme from "./themes/DefaultTheme";
 import { LexicalComposer } from "@lexical/react/LexicalComposer";
 import { RichTextPlugin } from "@lexical/react/LexicalRichTextPlugin";
 import { ContentEditable } from "@lexical/react/LexicalContentEditable";
@@ -28,7 +28,7 @@ function Placeholder() {
 
 const editorConfig = {
   // The editor theme
-  theme: ExampleTheme,
+  theme: DefaultTheme,
   // Handling of errors during update
   onError(error) {
     throw error;
